+++ index.php
@@ -3,20 +3,19 @@
  * Home File
  *
  * @author      David Voglgsang
- * @version     1.1
+ * @version     1.2
  *
 */
 
 get_header();
 ?>
 <section id="home" <?php echo prefix_template::AddContainer(prefix_template::$template_container, true); ?>>
-  <h1><?php the_archive_title(); ?></h1>
-  <?php if (have_posts() ) : while (have_posts()) : the_post() ?>
-    <article>
-        <h2><?php the_title(); ?></h2>
-      <?php the_excerpt(); ?>
-    </article>
-    <?php get_sidebar(); ?>
-  <?php endwhile; endif; ?>
+  <div>
+    <?php if (have_posts() ) : while (have_posts()) : the_post();
+      $blog_type = get_post_meta(get_the_ID(), 'template_blog_type', true) ? get_post_meta(get_the_ID(), 'template_blog_type', true) : "default";
+      get_template_part('template_parts/' . get_post_type() . '_' . $blog_type);
+    endwhile; endif; ?>
+  </div>
+  <?php get_sidebar(); ?>
 </section>
 <?php get_footer(); ?>